+++ mrs_plugin/db_schema/mysql_rest_service_metadata.msm.project/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Database Schema Change Log for `mysql_rest_service_metadata`
 
+## 4.0.3
+
+- Changed the export format of the audit_log dump to valid JSON
+- Enabled the `delete_old_audit_log_entries` EVENT by default
+- Added `msm_instance_demoted` and `msm_instance_promoted` procedures
+
 ## 4.0.2
 
 - Changes to the router_general_log table